feat: make ansi the default dialect
