+++ sqllineage/__init__.py
@@ -31,4 +31,4 @@ STATIC_FOLDER = "build"
 DEFAULT_HOST = "localhost"
 DEFAULT_PORT = 5000
 SQLPARSE_DIALECT = "non-validating"
-DEFAULT_DIALECT = SQLPARSE_DIALECT
+DEFAULT_DIALECT = "ansi"
