+++ tests/core/test_exception.py
@@ -2,6 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
+from sqllineage import SQLPARSE_DIALECT
 from sqllineage.exceptions import (
     InvalidSyntaxException,
     SQLLineageException,
@@ -13,18 +14,18 @@ from sqllineage.runner import LineageRunner
 def test_select_without_table():
     with pytest.raises(SQLLineageException):
         LineageRunner("select * from where foo='bar'")._eval()
+    with pytest.raises(SQLLineageException):
+        LineageRunner("select * from where foo='bar'", dialect=SQLPARSE_DIALECT)._eval()
 
 
 def test_full_unparsable_query_in_sqlfluff():
     with pytest.raises(InvalidSyntaxException):
-        LineageRunner("WRONG SELECT FROM tab1", dialect="ansi")._eval()
+        LineageRunner("WRONG SELECT FROM tab1")._eval()
 
 
 def test_partial_unparsable_query_in_sqlfluff():
     with pytest.raises(InvalidSyntaxException):
-        LineageRunner(
-            "SELECT * FROM tab1 AS FULL FULL OUTER JOIN tab2", dialect="ansi"
-        )._eval()
+        LineageRunner("SELECT * FROM tab1 AS FULL FULL OUTER JOIN tab2")._eval()
 
 
 def test_partial_unparsable_query_in_sqlfluff_with_tsql_batch():
@@ -38,14 +39,12 @@ CROSS JOIN tab2 AS src2"""
 
 def test_unsupported_query_type_in_sqlfluff():
     with pytest.raises(UnsupportedStatementException):
-        LineageRunner(
-            "CREATE UNIQUE INDEX title_idx ON films (title)", dialect="ansi"
-        )._eval()
+        LineageRunner("CREATE UNIQUE INDEX title_idx ON films (title)")._eval()
 
 
 def test_deprecation_warning_in_sqlparse():
     with pytest.warns(DeprecationWarning):
-        LineageRunner("SELECT * FROM DUAL", dialect="non-validating")._eval()
+        LineageRunner("SELECT * FROM DUAL", dialect=SQLPARSE_DIALECT)._eval()
 
 
 def test_syntax_warning_no_semicolon_in_tsql():
@@ -63,5 +62,4 @@ def test_user_warning_enable_tsql_no_semicolon_with_other_dialect():
         LineageRunner(
             """SELECT * FROM foo;
 SELECT * FROM bar""",
-            dialect="ansi",
         )._eval()
