+++ tests/sql/column/test_column_select_column.py
@@ -189,3 +189,45 @@ FROM tab2"""
             )
         ],
     )
+
+
+def test_non_reserved_keyword_as_column_name():
+    sql = """INSERT INTO tab1
+SELECT host
+FROM tab2 a"""
+    assert_column_lineage_equal(
+        sql,
+        [
+            (
+                ColumnQualifierTuple("host", "tab2"),
+                ColumnQualifierTuple("host", "tab1"),
+            )
+        ],
+        test_sqlparse=False,
+    )
+
+
+def test_current_timestamp():
+    """
+    current_timestamp is a keyword since ANSI SQL-2016
+    sqlparse cannot produce the correct AST for this case
+    """
+    sql = """INSERT INTO tab1
+SELECT current_timestamp as col1,
+       col2,
+       col3
+FROM tab2"""
+    assert_column_lineage_equal(
+        sql,
+        [
+            (
+                ColumnQualifierTuple("col2", "tab2"),
+                ColumnQualifierTuple("col2", "tab1"),
+            ),
+            (
+                ColumnQualifierTuple("col3", "tab2"),
+                ColumnQualifierTuple("col3", "tab1"),
+            ),
+        ],
+        test_sqlparse=False,
+    )
