+++ tests/sql/column/test_column_select_expression.py
@@ -131,3 +131,21 @@ FROM tab2 a"""
             ),
         ],
     )
+
+
+def test_subquery_expression_without_source_table():
+    assert_column_lineage_equal(
+        """INSERT INTO foo
+SELECT (SELECT col1 + col2 AS result) AS sum_result
+FROM bar""",
+        [
+            (
+                ColumnQualifierTuple("col1", "bar"),
+                ColumnQualifierTuple("sum_result", "foo"),
+            ),
+            (
+                ColumnQualifierTuple("col2", "bar"),
+                ColumnQualifierTuple("sum_result", "foo"),
+            ),
+        ],
+    )
