+++ tests/sql/column/test_metadata_unqualified_column.py
@@ -1,10 +1,10 @@
 import pytest
-from tests.helpers import assert_column_lineage_equal, generate_metadata_providers
 
 from sqllineage.core.metadata_provider import MetaDataProvider
 from sqllineage.exceptions import InvalidSyntaxException
 from sqllineage.runner import LineageRunner
 from sqllineage.utils.entities import ColumnQualifierTuple
+from ...helpers import assert_column_lineage_equal, generate_metadata_providers
 
 
 providers = generate_metadata_providers(
@@ -278,7 +278,7 @@ left join (select pk, p, q as z from db3.table3) t3 on t2.id = t3.pk
 
 @pytest.mark.parametrize("provider", providers)
 def test_sqlparse_exception(provider: MetaDataProvider):
-    sql = """insert into table db.tbl
+    sql = """insert into db.tbl
 select id
 from db1.table1 t1
 join db2.table2 t2 on t1.id = t2.id
