+++ tests/sql/column/test_metadata_wildcard.py
@@ -1,8 +1,8 @@
 import pytest
-from tests.helpers import assert_column_lineage_equal, generate_metadata_providers
 
 from sqllineage.core.metadata_provider import MetaDataProvider
 from sqllineage.utils.entities import ColumnQualifierTuple
+from ...helpers import assert_column_lineage_equal, generate_metadata_providers
 
 
 providers = generate_metadata_providers(
