--- tests/sql/sqlfluff_only/__init__.py
@@ -1,4 +1,0 @@
-"""
-This module includes all test cases that's supported by sqlfluff as parser only.
-By the time we deprecate sqlparse, we shall move these test cases back to tests.sql module
-"""
