+++ tests/sql/table/test_other_without_lineage_dialect_specific.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ....helpers import assert_table_lineage_equal
+from ...helpers import assert_table_lineage_equal
 
 
 @pytest.mark.parametrize("dialect", ["exasol", "mysql", "teradata"])
