+++ tests/sql/table/test_path_dialect_specific.py
@@ -1,7 +1,7 @@
 import pytest
 
 from sqllineage.core.models import Path
-from ....helpers import assert_table_lineage_equal
+from ...helpers import assert_table_lineage_equal
 
 
 @pytest.mark.parametrize("dialect", ["postgres", "redshift"])
